Remove factories, parser and runner
Added initial unit tests
