+++ .usher.yml
@@ -1,37 +1,20 @@
-build:
-  cmd: docker build
-  force_rm: true
-  tags:
-    - registry: &REGISTRY docker-registry.dun.fh
-      image: &IMAGE_NAME findmypast/usher
-      tag: <%=version%>
+vars:
+  registry: docker-registry.dun.fh
+  image: findmypast/usher
 
-push:
-  cmd: docker push
-  registry: *REGISTRY
-  image: *IMAGE_NAME
-  tag: <%=version%>
+tasks:
+  build:
+    cmd: docker build --force-rm -t <%=registry%>/<%=image%>:<%=version%> .
 
-test:
-  cmd: docker run
-  name: usher-test-runner
-  remove: true
-  target:
-    registry: *REGISTRY
-    image: *IMAGE_NAME
-    tag: <%=version%>
-  container_command: npm test
+  push:
+    cmd: docker push <%=registry%>/<%=image%>:<%=version%>
 
-publish:
-  cmd: docker run
-  name: usher-publisher
-  environment:
-    - NPM_USER=<%=user%>
-    - NPM_PASSWORD=<%=password%>
-    - NPM_EMAIL=<%=email%>
-  remove: true
-  target:
-    registry: *REGISTRY
-    image: *IMAGE_NAME
-    tag: <%=version%>
-  container_command: npm run publish-to-npm
+  test:
+    cmd: docker run --name usher-test-runner --rm <%=registry%>/<%=image%>:<%=version%>  npm test
+
+  publish:
+    cmd: docker run --name usher-publisher --rm <%=registry%>/<%=image%>:<%=version%> npm run publish-to-npm
+    environment:
+      - NPM_USER=<%=user%>
+      - NPM_PASSWORD=<%=password%>
+      - NPM_EMAIL=<%=email%>
