--- src/commands/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  init: require('./init'),
-  run: require('./run')  
-}
