--- src/factories/docker/index.js
@@ -1,6 +1,0 @@
-module.exports = {
-  build: require('./build'),
-  pull: require('./pull'),
-  push: require('./push'),
-  run: require('./run')
-}
