--- src/factories/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  docker: require('./docker')
-}
