--- src/templates/prompt.default.js
@@ -1,76 +1,0 @@
-'use strict';
-
-module.exports = [
-  {
-    name: "buildDockerRegistry",
-    message: "The docker registry URL",
-    validate: (value) => {
-      return new Promise( (resolve, reject) => {
-        if(value === '')
-            return resolve("buildDockerRegistry cannot be blank");
-        return resolve(true);
-      });
-    }
-  },
-  {
-    name: "buildDockerImage",
-    message: "The docker image, e.g. findmypast/usher",
-    validate: (value) => {
-      return new Promise( (resolve, reject) => {
-        if(value === '')
-            return resolve("buildDockerImage cannot be blank");
-        return resolve(true);
-      });
-    }
-  },
-  {
-    name: "projectName",
-    message: "The name of your project, e.g. usher",
-    validate: (value) => {
-      return new Promise( (resolve, reject) => {
-        if(value.search(/\s/g) !== -1)
-          return resolve("projectName cannot contain spaces");
-        if(value === '')
-            return resolve("projectName cannot be blank");
-        return resolve(true);
-      });
-    }
-  },
-  {
-    name: "clientPort",
-    message: "Port the service runs on locally within the container, e.g. 8080",
-    validate: (value) => {
-      return new Promise( (resolve, reject) => {
-        if(value.search(/[^\d]/g) !== -1)
-          return resolve("clientPort must be numeric");
-        if(value === '')
-            return resolve("clientPort cannot be blank");
-        return resolve(true);
-      });
-    }
-  },
-  {
-    name: "liveHostPort",
-    message: "Port the service should run on, on the host machine, e.g. 80",
-    validate: (value) => {
-      return new Promise( (resolve, reject) => {
-        if(value.search(/[^\d]/g) !== -1)
-          return resolve("liveHostPort must be numeric");
-        if(value === '')
-            return resolve("liveHostPort cannot be blank");
-        return resolve(true);
-      });
-    }
-  },
-  {
-    name: "remoteDockerAddress",
-    message: "Full URL to remote docker daemon, e.g. tcp://example.com:12345",
-    validate: (value) => {
-      return new Promise( (resolve, reject) => {
-        if(value === '')
-            return resolve("remoteDockerAddress cannot be blank");
-        return resolve(true);
-      });
-    }
-  },
-];
