--- src/templates/template.default.yml
@@ -1,56 +1,0 @@
-build:
-  cmd: docker build
-  force_rm: true
-  tags:
-    - registry: &REGISTRY {{buildDockerRegistry}}
-      image: &IMAGE_NAME {{buildDockerImage}}
-      tag: <%=version%>
-
-push:
-  cmd: docker push
-  registry: *REGISTRY
-  image: *IMAGE_NAME
-  tag: <%=version%>
-
-test:
-  cmd: docker run
-  name: {{projectName}}-test-runner
-  remove: true
-  target: &TARGET
-    registry: *REGISTRY
-    image: *IMAGE_NAME
-    tag: <%=version%>
-  container_command: {{testCommand}}
-
-local:
-  - cmd: docker rm -f {{projectName}}-local
-    ignore_errors:
-        - 1
-  - cmd: docker run
-    name: {{projectName}}-local
-    memory: 256M
-    publish:
-      - client: {{clientPort}}
-        host: {{clientPort}}
-    detach: true
-    target: *TARGET
-
-deploy:
-  - cmd: docker -H {{remoteDockerAddress}} rm -f {{projectName}}
-    ignore_errors:
-      - 1
-  - cmd: docker run
-    name: {{projectName}}
-    memory: 256M
-    publish:
-      - client: {{clientPort}}
-        host: {{liveHostPort}}
-    environment:
-      - SERVICE_NAME={{projectName}}
-      - SERVICE_TAGS=production
-      - SERVICE_CHECK_HTTP=/health
-      - SERVICE_CHECK_INTERVAL=60s
-    restart: always
-    detach: true
-    host: {{remoteDockerAddress}}
-    target: *TARGET
