--- src/templates/template.npm.yml
@@ -1,37 +1,0 @@
-build:
-  cmd: docker build
-  force_rm: {{buildForceRm}}
-  tags:
-    - registry: &REGISTRY {{buildDockerRegistry}}
-      image: &IMAGE_NAME {{buildDockerImage}}
-      tag: <%=version%>
-
-push:
-  cmd: docker push
-  registry: *REGISTRY
-  image: *IMAGE_NAME
-  tag: <%=version%>
-
-test:
-  cmd: docker run
-  name: {{buildDockerImage}}-test-runner
-  remove: true
-  target:
-    registry: *REGISTRY
-    image: *IMAGE_NAME
-    tag: <%=version%>
-  container_command: {{testCommand}}
-
-publish:
-  cmd: docker run
-  name: {{buildDockerImage}}-publisher
-  environment:
-    - NPM_USER=<%=user%>
-    - NPM_PASSWORD=<%=password%>
-    - NPM_EMAIL=<%=email%>
-  remove: true
-  target:
-    registry: *REGISTRY
-    image: *IMAGE_NAME
-    tag: <%=version%>
-  container_command: {{publishCommand}}
