--- test/integration/templates/prompt.usher.js
@@ -1,22 +1,0 @@
-module.exports = [
-  {
-    "name": "fastTestCommand",
-    "message": "The command to execute for the 'fast test' task"
-  },
-  {
-    "name": "buildDockerCommand",
-    "message": "The command to execute for the 'build' task"
-  },
-  {
-    "name": "buildForceRm",
-    "message": "The force_rm flag for the 'build' task"
-  },
-  {
-    "name": "buildDockerRegistry",
-    "message": "The docker registry URL for the 'build' task"
-  },
-  {
-    "name": "buildDockerImage",
-    "message": "The docker image for the 'build' task"
-  }
-];
