--- test/integration/templates/template.usher.yml
@@ -1,10 +1,0 @@
-fast_tests:
-  cmd: {{fastTestCommand}}
-
-build:
-  cmd: {{buildDockerCommand}}
-  force_rm: {{buildForceRm}}
-  tags:
-    - registry: &REGISTRY {{buildDockerRegistry}}
-      image: &IMAGE_NAME {{buildDockerImage}}
-      tag: <%=version%>
