--- test/integration/test.usher.yml
@@ -1,31 +1,0 @@
-fast_tests:
-  cmd: npm test
-
-build_prod:
-  cmd: docker build
-  force_rm: true
-  tags:
-    - registry: &REGISTRY docker-registry.dun.fh
-      image: &IMAGE_NAME findmypast/usher
-      tag: latest
-    - registry: *REGISTRY
-      image: *IMAGE_NAME
-      tag: <%=version%>
-
-run_prod:
-  cmd: docker run
-  environment:
-    - node-env=prod
-    - anotherenv=something
-  memory: 256M
-  name: usher
-  publish:
-    - client: 80
-      host: 80
-    - client: 8081
-      host: 8080
-  target:
-    registry: *REGISTRY
-    image: *IMAGE_NAME
-    tag: test
-  container_command: npm test
